+++ src/samples/sample-piral-core/src/pilets/index.ts
@@ -6,6 +6,7 @@ import { SearchPilet } from './search';
 import { FormPilet } from './form';
 import { VuePilet } from './vuePilet';
 import { NgPilet } from './ngPilet';
+import { HyperappPilet } from './hyperappPilet';
 import { ContainerPilet } from './container';
 
 /**
@@ -26,5 +27,6 @@ export const availablePilets = [
   FormPilet,
   VuePilet,
   NgPilet,
+  HyperappPilet,
   ContainerPilet,
 ];