+++ src/samples/sample-piral-core/src/types/index.ts
@@ -1,5 +1,6 @@
 import { PiralApi } from 'piral-core';
 import { PiralNgApi } from 'piral-ng';
 import { PiralVueApi } from 'piral-vue';
+import { PiralHyperappApi } from 'piral-hyperapp';
 
-export type SampleApi = PiralApi<PiralVueApi & PiralNgApi>;
+export type SampleApi = PiralApi<PiralVueApi & PiralNgApi & PiralHyperappApi>;